+++ dev/main.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-# This script is used to locally start the integration between SDP and AnchorPlatform for the SEP-24 deposit flow, needed for registering users.
+# This script is used to locally start the SDP environment for wallet registration and payment processing.
 set -eu
 
 export DIVIDER="----------------------------------------"
@@ -52,7 +52,7 @@ echo $DIVIDER
 
 # Check if "--delete_pv" is passed as a parameter
 if [[ " $@ " =~ " --delete_pv " ]]; then
-    echo "====> 👀 deleting persistent volumes sdp-multi-tenant_postgres-ap-db sdp-multi-tenant_postgres-db"
+    echo "====> 👀 deleting persistent volumes sdp-multi-tenant_postgres-db"
     
     # Function to delete volume if it exists
     delete_volume() {
@@ -66,7 +66,6 @@ if [[ " $@ " =~ " --delete_pv " ]]; then
     }
 
     # Delete volumes
-    delete_volume "sdp-multi-tenant_postgres-ap-db"
     delete_volume "sdp-multi-tenant_postgres-db"
 fi